Split faneffects.h into logical units
